+++ apps/trackers/tests/conftest.py
@@ -164,6 +164,13 @@ def jira_vulnissuetype_fields_setup_without_versions():
         ],
     ).save()
 
+    JiraProjectFields(
+        project_key="FOOPROJECT",
+        field_id="customfield_12325940",
+        field_name="Update Stream",
+        allowed_values=[],
+    ).save()
+
 
 @pytest.fixture()
 def setup_vulnerability_issue_type_fields() -> None: