+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Adjust Jira tracker sync to use `Update Stream` field and `ps_component` label (OSIDB-3830)
 
 ## [5.6.0] - 2026-03-03
 ### Added