+++ docs/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Serializer grant access to users with at least one reading LDAP group instead of enforcing all (OSIDB-4826)
+- Adapt Jira integration for Atlassian Cloud (OSIDB-4800)
+
+### Removed
+- Deprecate `team_id` and `group_key` fields from Flaw model and API (OSIDB-4540)
 
 ## [5.6.0] - 2026-03-03
 ### Added